requirements.txt
Flask==2.3.2
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
Flask-SQLAlchemy==3.0.3
email-validator==1.3.1
